music/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
fastapi==0.109.0
greenlet==3.0.3
h11==0.14.0
idna==3.6
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
PyMySQL==1.1.0
python-dotenv==1.0.0
requests==2.31.0
setuptools==68.2.2
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.35.1
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.25.0
wheel==0.41.2
